Fix hieraconf and pathing
The hieraconf host setting is not available on AIO because beaker has
moved to using Puppet setting helpers instead, so
`host.puppet['hiera_config']` is the prefered method.

The default paths are correctly configured for pe/foss <4, but not for
AIO unless the host type is AIO, so we have to work around that.

Updated the non *_on() method to run on the default node, as is the
normal pattern for such methods.