Redo collections. Use beaker's built-in code more.
The variable `PUPPET_INSTALL_TYPE` was originally introduced to control
selecting the PE/foss installation code, and later was changed to
allow the puppet-agent install code. It was incorrectly updated in 0.8.0
to also control collections however beaker already controls collections
via `BEAKER_PUPPET_COLLECTION` so this was not need (and broke beaker's
usage.)

`PUPPET_INSTALL_TYPE` still can be used to install 3.x foss instead of
puppet-agent, but that is deprecated.

Other variables in beaker also cover the functionality provided by some
in BPIH.