(maint) Install dev packages using newer utils
This commit updates how we install packages when given a sha. We assume
that if you're not hitting 'latest' or a specific version, you want to
install from our development repos, which are internal only. This
behavior is triggered when a user sets either `BEAKER_PUPPET_AGENT_SHA`
or `PUPPET_AGENT_SHA`. Even if these are versions (rather than shas),
you will still be able to install packages as long as they are present
on builds.delivery.puppetlabs.net.